+++ include/ui/tileseteditor.h
@@ -37,8 +37,9 @@ class TilesetEditor : public QMainWindow
 public:
     explicit TilesetEditor(Project*, Map*, QWidget *parent = nullptr);
     ~TilesetEditor();
-    void setMap(Map*);
-    void setTilesets(QString, QString);
+    void update(Map *map, QString primaryTilsetLabel, QString secondaryTilesetLabel);
+    void updateMap(Map *map);
+    void updateTilesets(QString primaryTilsetLabel, QString secondaryTilesetLabel);
     bool selectMetatile(uint16_t metatileId);
 
 private slots:
@@ -92,17 +93,25 @@ private slots:
     void on_actionImport_Secondary_Metatiles_triggered();
 
 private:
-    void init(Project*, Map*);
-    void closeEvent(QCloseEvent*);
-    void initMetatileSelector(Map*);
+    void initUi();
+    void setMetatileBehaviors();
+    void setMetatileLayersUi();
+    void setVersionSpecificUi();
+    void setMetatileLabelValidator();
+    void initMetatileSelector();
     void initTileSelector();
     void initSelectedTileItem();
     void initMetatileLayersItem();
+    void restoreWindowState();
+    void initMetatileHistory();
+    void setTilesets(QString primaryTilesetLabel, QString secondaryTilesetLabel);
+    void reset();
     void drawSelectedTiles();
     void importTilesetTiles(Tileset*, bool);
     void importTilesetMetatiles(Tileset*, bool);
     void refresh();
     void saveMetatileLabel();
+    void closeEvent(QCloseEvent*);
     Ui::TilesetEditor *ui;
     History<MetatileHistoryItem*> metatileHistory;
     TilesetEditorMetatileSelector *metatileSelector = nullptr;
@@ -110,6 +119,7 @@ private:
     MetatileLayersItem *metatileLayersItem = nullptr;
     PaletteEditor *paletteEditor = nullptr;
     Project *project = nullptr;
+    Map *map = nullptr;
     Metatile *metatile = nullptr;
     int paletteId;
     bool tileXFlip;